+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 - Removed `data::v2::stocks` module alias
 - Switched from using `test-env-log` to `test-log`
 - Bumped minimum supported Rust version to `1.56`
+- Added `async-trait` dependency in version `0.1.51`
 - Bumped `websocket-util` dependency to `0.10.1`
 - Bumped `tokio-tungstenite` dependency to `0.16`